implement setup helper hierarchy
set up directory structure to support multiple setup files changed
from:
  spec/support/setup_utils.rb
to:
  spec/support/setup/base.rb

This structure makes it easy to add setup helpers for other classes,
e.g., spec/support/setup/queue_manager.rb
