+++ spec/spec_helper.rb
@@ -1,3 +1,3 @@
 require_relative '../lib/git_transactor'
 require_relative '../test_support/helpers'
-require_relative './support/setup_utils'
+require_relative './support/setup/base'
